+++ frontend/js/js.required.ts
@@ -1,0 +1,14 @@
+export const required = {
+    "Index": [
+        "newButton.js",
+        "uploadButton.js",
+    ],
+    "Editor": [
+        "newButton.js",
+        "saveButton.js",
+        "uploadButton.js",
+        "imageButton.js",
+        "textViewer.js",
+        "titleContainer.js"
+    ],
+}