+++ frontend/js/public/newButton.js
@@ -1,0 +1,9 @@
+$("#new-button").on("click", async () => {
+  const response = await fetch("/new", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+  });
+  console.log(response);
+});